pycharm compatible.
need to place groups 1&2 in same dir as program in pycharm. install scipy, numpy, and matplotlib libraries
to install libraries: in open project with opened file: 
note: can copy and paste the code from the part4.ipynb file into a new python file in pycharm and run the program, too. 

 1.   ctrl-alt-s
 2.   click "project:projet name"
 3.   click project interperter
  4.  double-  click pip
  5.   search numpy from the top bar
   6.  click on numpy
  7.  click install package button
8. do steps 5-7 for each library